backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-jose
typing_extensions